Updated credentials handling for new aws sdk.
AWS credentials should generally not be specificed in code. There are
various ways in which the SDK can discover the credentials.

AWS SDK provides a default provider chain that does the right
thing. (see
http://docs.aws.amazon.com/AWSSdkDocsJava/latest/DeveloperGuide/credentials.html
)

This change updates the credentials handling in clj-aws-s3 such that if
nil is passed as the creds to any call then the AWS defaults will be
used. Alternatively you can explicitly supply a provider, or for
backwards compatibility supply the access-key-id and secret-access-key.

Ideally we would make cred optional in the API, but because of varargs
in various fn signatures that would make backwards compatibility
difficult, so we allow for supplying nil as the credentials.

We also use the strictly correct names for the two identifiers used by
AWS - access-key-id and secret-access-key. The old names are supported
for backwards compatibility.